Displayed the skill cooldowns on the battle screen.
Worked on skills, and modified skill class to take in level and calculate the distance and base damage of a skill based on level (implementation not done). Currently randomly generating enemy skill levels since they're not stored anywhere. Drew the skill cooldowns and added a set multiline to MenuText so that the cooldowns are evenly spaced with the radio btns and skill name. Also changed GameView so that a BattleView would only be instantiated when there is a battle.