Write getters for the information needed by the shop and the anomalous index
Added a lot of getters to the presenter and som more methods in utils to help get information from json files. Also added descriptions for each item and skill in their respective files, and used then to test the new presenter methods. Also added a get enemy image that returns a bitmap in the presenter, to be used to get the entity images in the anomalous index. Also added the icon for the anomalous index.